Fix wrong ParameterValueSet handling #390
- For each scenario run one parameter value set is now used
- Previous behavior was to apply all value sets to one single run

+++ EnvironmentSimulator/Libraries/esminiLib/esminiLib.cpp
@@ -498,12 +498,12 @@ extern "C"
 
 	SE_DLL_API int SE_GetNumberOfPermutations()
 	{
-		return OSCParameterDistribution::Inst().GetNumPermutations();
+		return static_cast<int>(OSCParameterDistribution::Inst().GetNumPermutations());
 	}
 
 	SE_DLL_API int SE_SelectPermutation(int index)
 	{
-		return OSCParameterDistribution::Inst().SetRequestedIndex(index);
+		return OSCParameterDistribution::Inst().SetRequestedIndex(static_cast<unsigned int>(index));
 	}
 
 	SE_DLL_API int SE_GetPermutationIndex()
